fixed a few "return false"'s -> "return true" updated docs a little

+++ src/api/c++/include/sqlrelay/private/sqlrcursor.h
@@ -98,9 +98,9 @@
 
 		// bind variables
 		bindvar		inbindvars[MAXVAR];
-		int		inbindcount;
+		unsigned short	inbindcount;
 		bindvar		outbindvars[MAXVAR];
-		int		outbindcount;
+		unsigned short	outbindcount;
 		int		validatebinds;
 
 		// result set
